feat: implement spotlight feature with global shortcut and modal
